refactor(D-168 A2): extract identity_access + stream_serialization from routes.py (verbatim)
- identity_access.py: session/thread/conversation-id resolution + JWT decode +
  admin payload/state/access guards + safe assistant error
- stream_serialization.py: NDJSON stream frames + ISS-056 human-readable
  extractor (_HUMAN_RESPONSE_FIELDS) + telemetry evidence line
- routes.py re-imports all 20 moved names (noqa: F401 re-exports)
- manifest: API_SOURCE_FILES += identity_access.py, stream_serialization.py

routes.py: 2,749 -> 2,465 lines. Gates green; 533 affected tests pass
(2 failures are the pre-existing CI-deselected pair).
